refactor: move complete test for pipenv into new integration test
