+++ .gitignore
@@ -152,3 +152,5 @@ Thumbs.db
 
 
 # End of https://www.toptal.com/developers/gitignore/api/python
+
+.coverage*
