+++ tests/integration/Pipfile
@@ -1,0 +1,12 @@
+[[source]]
+url = "https://pypi.python.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+texttable = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.8"
